+++ src/components/About.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const About = () => {
+   return(
+        <div className="card">
+            <h1 className="card-about">Todo List was created for Norma Moguel Class Ibis React February 2025</h1>
+        </div>
+
+  );
+}
+export default About;
