+++ src/components/Home.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Home =() => {
+   return(
+   <div className="card">
+     <h1 className="card-home">Wellcome TodoList</h1>
+     
+   </div>
+)
+}
+export default Home;
